All Admin functions added

+++ src/main/java/com/mycompany/service/CommentService.java
@@ -9,8 +9,11 @@ import org.springframework.stereotype.Service;
 
 import com.mycompany.dao.ICommentFunctionDAO;
 import com.mycompany.dao.IPostFunctionDAO;
+import com.mycompany.dao.IRoleFunctionDAO;
 import com.mycompany.dao.IUserFunctionDAO;
 import com.mycompany.entity.Comment;
+import com.mycompany.entity.Role;
+import com.mycompany.entity.User;
 
 @Service("commentService")
 public class CommentService {
@@ -21,6 +24,9 @@ public class CommentService {
 	@Autowired
 	private IUserFunctionDAO userDao;
 	
+	@Autowired
+	private IRoleFunctionDAO roleDao;
+	
 	public void addComment(Comment comment) {
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		comment.setUser(userDao.findByUsername(auth.getName()));
@@ -29,7 +35,12 @@ public class CommentService {
 	}
 	
 	public void deleteComment(int id) {
-		commentDao.deleteById(id);
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		User user = userDao.findByUsername(auth.getName());
+		Role role = roleDao.findById(2).get();
+		if(commentDao.findById(id).get().getUser().getId() == user.getId() || user.getRoles().contains(role)) {
+			commentDao.deleteById(id);
+		}
 	}
 
 }
